Commit after fixing second issue where the uploader of image can only edit or delete the image

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -45,7 +45,7 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{imageId}/{title}")
+    @RequestMapping(value = "/images/image/{imageId}/{title}")
     public String showImage(@PathVariable("imageId") Integer imageId, Model model) {
         Image image = imageService.getImageByTitle(imageId);
         model.addAttribute("image", image);
@@ -92,9 +92,18 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, @RequestParam("userId") Integer userId, Model model) {
         Image image = imageService.getImage(imageId);
+        String error = "Only the owner of the image can edit the image";
+        Integer uploaderId = imageService.getUploaderId(imageId);
 
+        if(uploaderId != userId){
+            model.addAttribute("editError",error);
+            Image image2 = imageService.getImageByTitle(imageId);
+            model.addAttribute("image", image2);
+            model.addAttribute("tags", image2.getTags());
+            return "images/image";
+        }
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
@@ -139,8 +148,17 @@ public class ImageController {
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
-    @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    @RequestMapping(value = "/deleteImage", method = RequestMethod.POST)
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, @RequestParam("userId") Integer userId, Model model) {
+        String error = "Only the owner of the image can delete the image";
+        Integer uploaderId = imageService.getUploaderId(imageId);
+        if(uploaderId != userId){
+            model.addAttribute("deleteError",error);
+            Image image2 = imageService.getImageByTitle(imageId);
+            model.addAttribute("image", image2);
+            model.addAttribute("tags", image2.getTags());
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
